♻️ : coupon form, list 컴포넌트로 분리
